+++ backend/Currency/Domain/Errors/OutOfRangeError.cs
@@ -4,7 +4,7 @@ namespace backend.Currency.Domain.Errors;
 
 public class OutOfRangeError : Error
 {
-    public OutOfRangeError(long amount) 
+    public OutOfRangeError(decimal amount) 
         : base("Amount out of range")
     {
         WithMetadata("Amount", amount);
